# wisataku-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/destinasi": {
      "get": {
        "operationId": "DestinasiController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daftar destinasi berhasil diambil"
          }
        },
        "summary": "Menampilkan daftar destinasi wisata",
        "tags": [
          "Destinasi"
        ]
      },
      "post": {
        "operationId": "DestinasiController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDestinasiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Destinasi berhasil dibuat"
          },
          "400": {
            "description": "Data tidak valid"
          }
        },
        "summary": "Menambahkan destinasi baru (khusus admin)",
        "tags": [
          "Destinasi"
        ]
      }
    },
    "/destinasi/{id}": {
      "get": {
        "operationId": "DestinasiController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail destinasi berhasil diambil"
          },
          "404": {
            "description": "Destinasi tidak ditemukan"
          }
        },
        "summary": "Menampilkan detail destinasi berdasarkan ID",
        "tags": [
          "Destinasi"
        ]
      },
      "patch": {
        "operationId": "DestinasiController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDestinasiDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Destinasi berhasil diperbarui"
          },
          "400": {
            "description": "Data tidak valid"
          },
          "404": {
            "description": "Destinasi tidak ditemukan"
          }
        },
        "summary": "Memperbarui data destinasi berdasarkan ID",
        "tags": [
          "Destinasi"
        ]
      },
      "delete": {
        "operationId": "DestinasiController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Destinasi berhasil dihapus"
          },
          "404": {
            "description": "Destinasi tidak ditemukan"
          }
        },
        "summary": "Menghapus destinasi berdasarkan ID",
        "tags": [
          "Destinasi"
        ]
      }
    }
  },
  "info": {
    "title": "WisataKu API",
    "description": "API untuk platform informasi dan reservasi destinasi wisata",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateDestinasiDto": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "example": "Pantai Kuta Mandalika"
          },
          "kategori": {
            "type": "string",
            "example": "Pantai"
          },
          "hargaTiket": {
            "type": "number",
            "example": 15000
          }
        },
        "required": [
          "nama",
          "kategori",
          "hargaTiket"
        ]
      },
      "UpdateDestinasiDto": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "example": "Pantai Kuta Mandalika"
          },
          "kategori": {
            "type": "string",
            "example": "Pantai"
          },
          "hargaTiket": {
            "type": "number",
            "example": 15000
          }
        }
      }
    }
  }
}