# wisataku-api/src/destinasi/dto/update-destinasi.dto.ts
import { PartialType } from '@nestjs/swagger';
import { CreateDestinasiDto } from './create-destinasi.dto.js';

// PartialType membuat seluruh field opsional — cocok untuk PATCH
export class UpdateDestinasiDto extends PartialType(CreateDestinasiDto) {}